\name{closeDatafiles}
\alias{closeDatafiles}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Close all open datafile handles
}
\description{
Iterates through all EPOCObjects and closes any open datafile handles.
}
\usage{
closeDatafiles(.Object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{.Object}{
Controller object.
}
}
\details{}
\value{}
\references{}
\author{
Troy Robertson
}
\note{}
\seealso{
\code{\link{EPOC}, \linkS4class{Controller}}
}
\examples{
## Close all datafiles
# closeDatafiles(controller)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
