\name{dayFromDate}
\alias{dayFromDate}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Return day of the year.
}
\description{
Takes a numeric day and a numeric month value and returns the Julian day of the year.
}
\usage{
dayFromDate(day = "numeric", month = "numeric")
}
\arguments{
  \item{day}{
Day of the month
}
  \item{month}{
Month of the year
}
}
\details{
}
\value{
}
\references{
}
\author{
Andrew Constable
}
\note{}
\seealso{
	\code{\link{EPOC}}
}
\examples{
## Return day of year
dayFromDate(day=15, month = 3)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
