\name{getActionMatForPeriod}
\alias{getActionMatForPeriod}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Calendar method
}
\description{
Return the action matrix for the period index specified.
}
\usage{
getActionMatForPeriod(.Object, periodNum)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{.Object}{
Calendar object
}
  \item{periodNum}{
Period number
}
}
\details{}
\value{}
\references{}
\author{
Andrew Constable, Troy Robertson
}
\note{}

\seealso{
	\code{\linkS4class{Calendar}, \linkS4class{Period}, \link{getPeriod}, \link{getInfoForPeriod}, \link{createCalendar}, \link{printCalendar}}
}
\examples{
## Return action matrix for period 3
# getActionMatForPeriod(calendar, 3)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
