\name{getAttributeNames}
\alias{getAttributeNames}
\title{
EPOCObject methods
}
\description{
Return a list containing all EPOC Attribute names from the epocAttributes slot list.
}
\usage{
getAttributeNames(.Object)
}
\arguments{
  \item{.Object}{
EPOCObject
}
}
\details{}
\value{
List of EPOC Attribute names
}
\references{}
\author{
Troy Robertson
}
\note{}
\seealso{
	\code{\linkS4class{EPOCObject}, \link{getAttribute}, \link{setAttribute}}
}
\examples{
## Return the list of attribute items
# getAttributeNames(epocObject)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
