\name{getBirthday}
\alias{getBirthday}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Element methods
}
\description{
Return the element's Julian birthday which represents the start time of an annual cycle for that element.
}
\usage{
getBirthday(.Object)
}
\arguments{
  \item{.Object}{
Element
}
}
\details{}
\value{
Returns an integer from 1 - 365.
}
\references{}
\author{
Troy Robertson
}
\note{}
\seealso{
	\code{\linkS4class{Element}, \linkS4class{EPOCObject}}
}
\examples{
## Get birthday
# bday <- getBirthday(element)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
