\name{getCalendar}
\alias{getCalendar}
\title{
Controller methods
}
\description{
Return the calendar object held by the controller
}
\usage{
getCalendar(.Object)
}
\arguments{
  \item{.Object}{
Controller object
}
}
\details{}
\value{
Calendar object
}
\references{}
\author{
Troy Robertson
}
\note{}
\seealso{
	\code{\linkS4class{Controller}, \linkS4class{Calendar}, \link{getUniverse}}
}
\examples{
## Get calendar object
# cont <- getCalendar(controller)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
