\name{getFullSignature}
\alias{getFullSignature}
\title{
EPOC Signature methods
}
\description{
Return a multi-line textual representation of this Signature object
}
\usage{
getFullSignature(.Object)
}
\arguments{
  \item{.Object}{
Signature object
}
}
\details{
}
\value{
return mutli-line string
}
\references{
}
\author{
Troy Robertson
}
\note{
}

\seealso{
	\code{\linkS4class{Universe}, \link{getSimpleSignature}, \link{getSignatureItem}, \link{setSignature}}
}
\examples{
## Get element signature and then get it's textual full signature
# sig <- getSignature(element)
# getFullSignature(sig)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
