\name{getPeriod}
\alias{getPeriod}
\title{
Calendar method
}
\description{
Return Period object for the period number specified.
}
\usage{
getPeriod(.Object, periodNum)
}
\arguments{
  \item{.Object}{
Calendar object
}
  \item{periodNum}{
Period number
}
}
\details{}
\value{
Period object
}
\references{}
\author{}
\note{}
\seealso{
	\code{\linkS4class{Calendar}, \linkS4class{Period}, \link{getInfoForPeriod}, \link{getActionMatForPeriod}, \link{createCalendar}, \link{printCalendar}}
}
\examples{
## Return 2nd period object
# getPeriod(calendar, 2)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
