\name{getPeriodActionMat}
\alias{getPeriodActionMat}
\title{
Period method
}
\description{
Return action matrix for this period object
}
\usage{
getPeriodActionMat(.Object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{.Object}{
Calendar object
}
}
\details{}
\value{
Action matrix for period object
}
\references{}
\author{
Troy Robertson
}
\note{}
\seealso{
	\code{\linkS4class{Period}, \link{getPeriod}, \link{getPeriodInfo}, \link{getPeriodElementTSData}}
}
\examples{
## Return period action matrix
# getPeriodActionMat(period)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
