\name{getPeriodInfo}
\alias{getPeriodInfo}
\title{
Period method
}
\description{
Return list containing period information for this period object.
}
\usage{
getPeriodInfo(.Object)
}
\arguments{
  \item{.Object}{
Period object
}
}
\details{}
\value{
Return a named list containing as:
list(Number, Day, KnifeEdge, YearPropn, PeriodStart, PeriodEnd)
}
\references{}
\author{
Troy Robertson
}
\note{}
\seealso{
	\code{\linkS4class{Period}, \link{getPeriod}, \link{getPeriodActionMat}, \link{getPeriodElementTSData}}
}
\examples{
## Return period information
# getPeriodInfo(period)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
