\name{getPolygons}
\alias{getPolygons}
\title{
Element methods
}
\description{
Return the polygons vector if it exists.  This a a vector of indexes to the universe spatial object polygons.
}
\usage{
getPolygons(.Object)
}
\arguments{
  \item{.Object}{
Element object
}
}
\details{}
\value{
Returns a vector of spatial polygon indexes.
}
\references{}
\author{
Troy Robertson
}
\note{}
\seealso{
	\code{\linkS4class{Element}, \linkS4class{EPOCObject}}
}
\examples{
## Get polygons
# getPolygons(element)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
