\name{getSignatureLine}
\alias{getSignatureLine}
\title{
EPOCObject method
}
\description{
Return a one-line textual signature for this object.
}
\usage{
getSignatureLine(.Object, display=FALSE)
}
\arguments{
  \item{.Object}{
EPOCObject
}
  \item{display}{
display	send string to stdout using epocMessage (default = FALSE)
}
}
\details{}
\value{
return one-line string
}
\references{
}
\author{ Troy Robertson }
\note{}
\seealso{
	\code{\linkS4class{EPOCObject}, \link{getSignature}, \link{getSignatureMulti}}
}
\examples{
## Return signature one liner
# getSignatureLine(element)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
