\name{getSimpleSignature}
\alias{getSimpleSignature}
\title{
EPOC Signature methods
}
\description{
Return this signature as a one line text.
}
\usage{
getSimpleSignature(.Object)
}
\arguments{
  \item{.Object}{
Signature object
}
}
\details{
}
\value{
returns a single line text string
}
\references{
}
\author{
Troy Robertson
}
\note{
}

\seealso{
	\code{\linkS4class{EPOCObject}, \linkS4class{Signature}, \link{getFullSignature}, \link{getSignatureItem}, \link{setSignature}}
}
\examples{
## Get text from Signature object sig
# sig <- getSignature(element)
# getSimpleSignature(sig)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
