\name{getSlotNames}
\alias{getSlotNames}
\title{
EPOCObject methods
}
\description{}
\usage{
getSlotNames(.Object)
}
\arguments{
  \item{.Object}{
EPOCObject
}
}
\details{}
\value{
A list of slot names
}
\references{}
\author{
Troy Robertson
}
\note{}
\seealso{
	\code{\linkS4class{EPOCObject}, \link{getSlot}, \link{setSlot}}
}
\examples{
## Get slot names
# getSlotNames(element)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
