\name{getUniverse}
\alias{getUniverse}
\title{
Controller methods
}
\description{
Return the universe object held by this controller
}
\usage{
getUniverse(.Object)
}
\arguments{
  \item{.Object}{
Controller object
}
}
\details{}
\value{
Universe object
}
\references{}
\author{
Troy Robertson
}
\note{}
\seealso{
	\code{\linkS4class{Controller}, \linkS4class{Calendar}, \link{getCalendar}}
}
\examples{
## Get universe object
# uni <- getUniverse(controller)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
