\name{initialiseElementTransitions}
\alias{initialiseElementTransitions}
\title{
Scenario methods
}
\description{
Method to call initialiseTransition methods for all elements in universe./cr
Called by runSimulation prior to starting each tria of the simulation.
}
\usage{
initialiseElementTransitions(.Object, universe = Universe)
}
\arguments{
  \item{.Object}{
Scenario object
}
  \item{universe}{
Universe object
}
}
\details{}
\value{}
\references{}
\author{
Troy Robertson
}
\note{}

\seealso{
	\code{\linkS4class{Scenario}, \linkS4class{Universe}, \linkS4class{EPOCObject}, link{initialiseElementScenarios}}
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
