\name{initialiseTransition}
\alias{initialiseTransition}
\title{
Element generic methods
}
\description{
Generic method to perform any pre-simulation initialisation of transition state data.
}
\usage{
initialiseTransition(.Object, universe="environment")
}
\arguments{
  \item{.Object}{
Element object.
}
\item{universe}{
Universe object.
}
}
\details{}
\value{}
\references{}
\author{
Troy Robertson
}
\note{}
\seealso{
	\code{\linkS4class{Element}, \linkS4class{EPOCObject}, link{initialiseReplicate}, link{updateState}, link{printState}}
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
