\name{parseInputData}
\alias{parseInputData}
\title{
EPOCObject methods
}
\description{
Parse objects input data that was sourced on initialisation.
}
\usage{
parseInputData(.Object)
}
\arguments{
  \item{.Object}{
EPOCObject
}
}
\details{
Parse input file data and insert into slot, attribute or signature as appropriate
}
\value{
Returns EPOCObject with data member changes.
}
\references{}
\author{ Troy Robertson }
\note{}
\seealso{
	\code{\linkS4class{EPOCObject}}
}
\examples{
## Parse data
# parseInputData(obj)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
