\name{testInputPaths}
\alias{testInputPaths}
\title{
Universe methods
}
\description{
Test all input data file paths to ensure they are valid files and return a list of failed paths.
}
\usage{
testInputPaths(.Object)
}
\arguments{
  \item{.Object}{
Universe object
}
}
\details{}
\value{
A list of invalid paths (zero length if all is well)
}
\references{}
\author{
Troy Robertson
}
\note{}
\seealso{
\code{\link{EPOC}, \linkS4class{Universe}, \linkS4class{Element}}
}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
