\name{updateState}
\alias{updateState}
\title{
Element generic methods
}
\description{
Generic method to perform update of element state data from transition state.
}
\usage{
updateState(.Object, universe = "environment")
}
\arguments{
  \item{.Object}{
Element object
}
  \item{universe}{
Universe object
}
}
\details{}
\value{}
\references{}
\author{
Troy Robertson
}
\note{}
\seealso{
	\code{\linkS4class{Element}, \linkS4class{EPOCObject}, link{initialiseReplicate}, link{initialiseTransition}, link{printState}}
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
